right queue integration
