+++ src/scheduler/CMakeLists.txt
@@ -8,6 +8,7 @@ set(LIB_FILES
         SnapshotServer.cpp
         SnapshotClient.cpp
         MpiContext.cpp
+        MpiThreadPool.cpp
         MpiWorldRegistry.cpp
         MpiWorld.cpp
         ${HEADERS}
