+++ src/scheduler/MpiThreadPool.cpp
@@ -1,8 +1,17 @@
-#include <threadPool.h>
+#include <faabric/scheduler/MpiThreadPool.h>
+#include <faabric/util/logging.h>
 
+namespace faabric::scheduler {
 MpiAsyncThreadPool::MpiAsyncThreadPool(int nThreads)
     : shutdown(false)
 {
+    faabric::util::getLogger()->debug("Starting an MpiAsyncThreadPool of size {}",
+                                      nThreads);
+
+    // Initialize job queue
+    localJobQueue = std::make_shared<MpiJobQueue>();
+
+    // Initialize thread pool
     for (int i = 0; i < nThreads; ++i) {
         threadPool.emplace_back(std::bind(&MpiAsyncThreadPool::entrypoint, this, i));
     }
@@ -10,16 +19,18 @@ MpiAsyncThreadPool::MpiAsyncThreadPool(int nThreads)
 
 MpiAsyncThreadPool::~MpiAsyncThreadPool()
 {
-    std::cout << "Shutting down MpiAsyncThreadPool" << std::endl;
+
+    faabric::util::getLogger()->debug("Shutting down MpiAsyncThreadPool");
     this->shutdown = true;
 
     // When the destructor is called, either some thread will have hit the
     // dequeue timeout and won't be joinable, or none has and all are joinable.
     // We wait untill all hit the timeout, saving an additional cond. variable.
-    for (auto& thread : threadPool)
+    for (auto& thread : threadPool) {
         if (thread.joinable()) {
             thread.join();
         }
+    }
 }
 
 void MpiAsyncThreadPool::awaitAsyncRequest(int reqId)
@@ -28,16 +39,20 @@ void MpiAsyncThreadPool::awaitAsyncRequest(int reqId)
 
     // Wait until the condition is signaled and the predicate is false.
     // Note that the same lock protects concurrent accesses to asyncReqMap
-    awakeCV.wait(lock, [reqId]{ 
-        return asyncReqMap.find(reqId) != asyncReqMap.end();
+    awakeCV.wait(lock, [this, reqId]{ 
+        return this->asyncReqMap.find(reqId) != this->asyncReqMap.end();
     });
 
     // Before giving up the lock, remove our request as it has already finished
     auto it = asyncReqMap.find(reqId);
     if (it == asyncReqMap.end()) {
-        std::cout << "Error: something went wrong." << std::endl;
+        throw std::runtime_error(fmt::format("Error: unrecognized reqId {}", reqId));
     }
     asyncReqMap.erase(it);
+}
+
+std::shared_ptr<MpiJobQueue> MpiAsyncThreadPool::getMpiJobQueue() {
+    return this->localJobQueue;
 }
 
 void MpiAsyncThreadPool::entrypoint(int i)
@@ -50,7 +65,7 @@ void MpiAsyncThreadPool::entrypoint(int i)
         // Note - we assume that if we hit the timeout it's time to shutdown
         // TODO -define the timeout in a constant
         try {
-            job = jobQueue.dequeue(5000);
+            job = getMpiJobQueue()->dequeue(5000);
         } catch (const faabric::util::QueueTimeoutException& e) {
             if (!this->shutdown) {
                 this->shutdown = true;
@@ -61,8 +76,6 @@ void MpiAsyncThreadPool::entrypoint(int i)
         int reqId = job.first;
         std::function<void (void)> func = job.second;
 
-        std::cout << "Executing on thread: " << i << std::endl;
-
         // Do the job without holding any locks
         func();
 
@@ -71,37 +84,12 @@ void MpiAsyncThreadPool::entrypoint(int i)
             faabric::util::UniqueLock lock(awakeMutex);
             auto it = asyncReqMap.insert(std::make_pair(reqId, 1));
             if (it.second == false) {
-                std::cout << "Error: reqId was already in the map!" << std::endl;
+                throw std::runtime_error(fmt::format("Error: reqId collision {}", reqId));
             }
         }
 
         // Notify that the work is done lock-free
         awakeCV.notify_all();
     }
-
 }
-
-/*
-int main()
-{
-    {
-        // Create two threads
-        MpiAsyncThreadPool p(2);
-
-        // Assign them 4 jobs
-        jobQueue.enqueue(std::make_pair(1, std::bind(silly, 1)));
-        jobQueue.enqueue(std::make_pair(2, std::bind(silly, 2)));
-        jobQueue.enqueue(std::make_pair(3, std::bind(silly, 3)));
-        jobQueue.enqueue(std::make_pair(4, std::bind(silly, 4)));
-
-        // Who is doing my job?
-        p.awaitAsyncRequest(1);
-        p.awaitAsyncRequest(2);
-        p.awaitAsyncRequest(3);
-        p.awaitAsyncRequest(4);
-
-        std::cout << "we get here" << std::endl;
-    }
-    return 0;
 }
-*/
